+++ tests/files/src/5372_trait_require.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace NS5372;
+
+class BaseClass {}
+
+interface FooInterface {}
+
+/**
+ * @require-extends BaseClass
+ */
+trait RequiresBase
+{
+    public function mustExtendBase(): void {}
+}
+
+/**
+ * @psalm-require-implements FooInterface
+ */
+trait RequiresInterface
+{
+    public function mustImplementInterface(): void {}
+}
+
+class ValidChild extends BaseClass
+{
+    use RequiresBase;
+}
+
+class ValidImplementation implements FooInterface
+{
+    use RequiresInterface;
+}
+
+class InvalidChild
+{
+    use RequiresBase;
+}
+
+class InvalidImplementation
+{
+    use RequiresInterface;
+}
+
+function test(ValidChild $child, ValidImplementation $impl): void
+{
+    $child->mustExtendBase();
+    $impl->mustImplementInterface();
+}